Split up alltests.py into separate unit test files.
* Put each test class in its own fie
* Re-enable the working tests in __init__.py
* Reduce imports to necessary used imports
